Lots of mods

+++ src/OLT.Email/OLT.Email/Configuration/OltSmtpServer.cs
@@ -3,7 +3,7 @@
     public class OltSmtpServer 
     {
         public virtual string Host { get; set; }
-        public virtual short? Port { get; set; }
+        public virtual int? Port { get; set; }
         public virtual bool DisableSsl { get; set; }
         public virtual OltSmtpCredentials Credentials { get; set; } = new OltSmtpCredentials();
     }
